// model/metric.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GraphTracks Analytics API
 * Bluesky Analytics API for GraphTracks
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * The metric to get stats for
 * @export
 * @enum {string}
 */

export const Metric = {
    Followers: 'followers',
    Likes: 'likes',
    Replies: 'replies',
    Reposts: 'reposts',
    Engagement: 'engagement'
} as const;

export type Metric = typeof Metric[keyof typeof Metric];